Initial commit with early codebase
This includes:
  * Client -- Basic class for handling client logic and function.
    * Includes dynamic module loading, allowing for different client
      frontends, macro systems, network stacks, and similar.
  * SDL2 Interface Module -- A stub for the main SDL2 interface.
  * Cfg -- Basic key=>value storage with file loading.

Only tested on Arch Linux w/ 4.13.12 kernel, but it shouldn't be too
difficult to have it running on MacOS and Windows. In the case of the
latter, it'll just be making the appropriate calls to `LoadLibrary`,
`GetProcAddress`, and various other friends. MacOS should be a breeze,
since it's POSIX.

Oh, and the entire codebase expects C++14, although C++11 could work
with minor changes.