Add ErrorDialog system and MacOS implementation
This simply creates a Cocoa dialog window that can be used to show errors. At the moment, it is used in handling `ModuleException`s that make it up to `main()` along with any unhandled exceptions (sans `.which()`, of course).

Also adjusted the SDL2 module to use ModuleException rather than SDL2::Exception. Internally, it should probably use its own exception classes, but, as far as throwing to main program, it should use ModuleException.